# kloggerx-server/docs/swagger.yaml
basePath: /api/v1
definitions:
  model.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: KloggerX 知识协作平台 API 文档
  title: KloggerX API
  version: "1.0"
paths:
...
  /admin/storage/usage:
    get:
      description: 管理员获取所有用户的存储使用情况
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 获取全局存储使用情况
      tags:
      - 系统管理
...
  /onlyoffice/callback/{docId}:
    post:
      consumes:
      - application/json
      description: 处理OnlyOffice服务器的回调请求
      parameters:
      - description: 文档ID
        in: path
        name: docId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: OnlyOffice回调
      tags:
      - 文档管理
  /onlyoffice/config:
    post:
      consumes:
      - application/json
      description: 获取文档的OnlyOffice编辑器配置
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 获取OnlyOffice编辑器配置
      tags:
      - 文档管理
  /onlyoffice/download/{docId}:
    get:
      description: 为OnlyOffice提供文档内容下载
      parameters:
      - description: 文档ID
        in: path
        name: docId
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: OnlyOffice文档下载
      tags:
      - 文档管理
  /onlyoffice/server-url:
    get:
      description: 返回OnlyOffice服务器地址
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 获取OnlyOffice服务器URL
      tags:
      - 文档管理
...
  /storage/stats:
    get:
      description: 获取系统存储统计信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 获取存储统计
      tags:
      - 文件
  /storage/usage:
    get:
      description: 获取当前用户的存储使用情况
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 获取存储使用情况
      tags:
      - 文件
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"